pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.0"
pydantic = "^2.11.7"
streamlit = "^1.46.1"
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
pydantic-settings = "^2.10.1"
python-jose = "^3.5.0"
requests = "^2.32.4"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
email-validator = "^2.2.0"
passlib = "^1.7.4"
greenlet = "^3.2.3"
bcrypt = "^4.3.0"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
